add tests and benchmarks packages
