+++ gob/receivers_test.go
@@ -3,27 +3,9 @@ package gob
 import (
 	"testing"
 
-	"github.com/doubledutch/mux"
+	"github.com/doubledutch/mux/tests"
 )
 
 func TestReceiver(t *testing.T) {
-	strCh := make(chan string, 1)
-	expected := "hello world"
-
-	pool := new(Pool)
-
-	strR := mux.NewReceiver(strCh, pool)
-	defer strR.Close()
-
-	enc := pool.NewBufferEncoder()
-	enc.Encode(expected)
-
-	if err := strR.Receive(enc.Bytes()); err != nil {
-		t.Fatal(err)
-	}
-
-	actual := <-strCh
-	if actual != expected {
-		t.Fatalf("actual '%s' != expected '%s'", actual, expected)
-	}
+	tests.Receiver(t, new(Pool))
 }
